removed OwSimProxy class, moving its meager content into OwAdapter
