+++ src/plexil-adapter/OwAdapter.cpp
@@ -7,7 +7,6 @@
 // OW
 #include "OwAdapter.h"
 #include "OwInterface.h"
-#include "OwSimProxy.h"
 #include "subscriber.h"
 
 // ROS
@@ -39,6 +38,111 @@ const Value Unknown;
 // An empty argument vector.
 static vector<Value> const EmptyArgs;
 
+
+//////////////////////// PLEXIL Lookup Support //////////////////////////////
+
+static void stubbed_lookup (const string& name, const string& value)
+{
+  // This warning is annoying.  Could parameterize it.
+  //  ROS_WARN("PLEXIL Adapter: Stubbed lookup of %s returning %s",
+  //           name.c_str(), value.c_str());
+}
+
+
+// NOTE: This macro, and the stub it implements, are temporary.
+#define STATE_STUB(name,val)                    \
+  if (state_name == #name) {                    \
+    stubbed_lookup (#name, #val);               \
+    value_out = val;                            \
+  }
+
+static bool lookup (const std::string& state_name,
+                    const std::vector<PLEXIL::Value>& args,
+                    PLEXIL::Value& value_out)
+{
+  bool retval = true;
+
+  // Stubbed mission and system parameters.  Many of these will eventually be
+  // obsolete.
+
+  STATE_STUB(TrenchLength, 10)
+  else STATE_STUB(TrenchGroundPosition, -0.155)
+  else STATE_STUB(TrenchWidth, 10)
+  else STATE_STUB(TrenchDepth, 2)
+  else STATE_STUB(TrenchPitch, 0)
+  else STATE_STUB(TrenchYaw, 0)
+  else STATE_STUB(TrenchStartX, 5)
+  else STATE_STUB(TrenchStartY, 10)
+  else STATE_STUB(TrenchStartZ, 0)
+  else STATE_STUB(TrenchDumpX, 0)
+  else STATE_STUB(TrenchDumpY, 0)
+  else STATE_STUB(TrenchDumpZ, 5)
+  else STATE_STUB(TrenchIdentified, true)
+  else STATE_STUB(TrenchTargetTimeout, 60)
+  else STATE_STUB(ExcavationTimeout, 10)
+  else STATE_STUB(ExcavationTimeout, 60)
+  else STATE_STUB(SampleGood, true)
+  else STATE_STUB(CollectAndTransferTimeout, 10)
+
+  else if (state_name == "TiltDegrees") {
+    value_out = OwInterface::instance()->getTilt();
+  }
+  else if (state_name == "PanDegrees") {
+    value_out = OwInterface::instance()->getPanDegrees();
+  }
+  else if (state_name == "PanVelocity") {
+    value_out = OwInterface::instance()->getPanVelocity();
+  }
+  else if (state_name == "TiltVelocity") {
+    value_out = OwInterface::instance()->getTiltVelocity();
+  }
+  else if (state_name == "HardTorqueLimitReached") {
+    string s;
+    args[0].getValue(s);
+    value_out = OwInterface::instance()->hardTorqueLimitReached(s);
+  }
+  else if (state_name == "SoftTorqueLimitReached") {
+    string s;
+    args[0].getValue(s);
+    value_out = OwInterface::instance()->softTorqueLimitReached(s);
+  }
+  else if (state_name == "Running") {
+    string operation;
+    args[0].getValue(operation);
+    value_out = OwInterface::instance()->running (operation);
+  }
+  else if (state_name == "StateOfCharge") {
+    value_out = OwInterface::instance()->getStateOfCharge();
+  }
+  else if (state_name == "RemainingUsefulLife") {
+    value_out = OwInterface::instance()->getRemainingUsefulLife();
+  }
+  else if (state_name == "BatteryTemperature") {
+    value_out = OwInterface::instance()->getBatteryTemperature();
+  }
+  else if (state_name == "GroundFound") {
+    value_out = OwInterface::instance()->groundFound();
+  }
+  else if (state_name == "GroundPosition") {
+    value_out = OwInterface::instance()->groundPosition();
+  }
+  // Faults
+  else if (state_name == "SystemFault") {
+    value_out = OwInterface::instance()->systemFault();
+  }
+  else if (state_name == "AntennaFault") {
+    value_out = OwInterface::instance()->antennaFault();
+  }
+  else if (state_name == "ArmFault") {
+    value_out = OwInterface::instance()->armFault();
+  }
+  else if (state_name == "PowerFault") {
+    value_out = OwInterface::instance()->powerFault();
+  }
+  else retval = false;
+
+  return retval;
+}
 
 //////////////////////////// Command Handling //////////////////////////////
 
@@ -343,13 +447,6 @@ bool OwAdapter::isStateSubscribed(const State& state) const
   return m_subscribedStates.find(state) != m_subscribedStates.end();
 }
 
-///////////////////////////// Domain Support //////////////////////////////////
-
-// We keep the domain interface as lightweight as possible!
-
-static OwSimProxy* TheSimProxy = 0;
-
-
 
 ///////////////////////////// Member functions //////////////////////////////////
 
@@ -358,13 +455,11 @@ OwAdapter::OwAdapter(AdapterExecInterface& execInterface,
                      const pugi::xml_node& configXml) :
   InterfaceAdapter(execInterface, configXml)
 {
-  TheSimProxy = new OwSimProxy();
   debugMsg("OwAdapter", " created.");
 }
 
 OwAdapter::~OwAdapter ()
 {
-  if (TheSimProxy) delete TheSimProxy;
 }
 
 bool OwAdapter::initialize()
@@ -384,7 +479,7 @@ bool OwAdapter::initialize()
   g_configuration->registerCommandHandler("tilt_antenna", tilt_antenna);
   g_configuration->registerCommandHandler("pan_antenna", pan_antenna);
   g_configuration->registerCommandHandler("take_picture", take_picture);
-  
+
   TheAdapter = this;
   setSubscriber (receiveBool);
   setSubscriber (receiveString);
@@ -435,7 +530,7 @@ void OwAdapter::lookupNow (const State& state, StateCacheEntry& entry)
 
   Value retval = Unknown;  // the value of the queried state
 
-  if (! TheSimProxy->lookup(state.name(), state.parameters(), retval)) {
+  if (! lookup(state.name(), state.parameters(), retval)) {
     ROS_ERROR("PLEXIL Adapter: Invalid lookup name: %s", state.name().c_str());
   }
   entry.update(retval);
